@@ -3,6 +3,7 @@ package main
 import (
 	"flag"
 	"os"
+	"fmt"
 
 	"github.com/go-kratos/kratos-layout/internal/conf"
 
@@ -13,20 +14,24 @@ import (
 	"github.com/go-kratos/kratos/v2/middleware/tracing"
 	"github.com/go-kratos/kratos/v2/transport/grpc"
 	"github.com/go-kratos/kratos/v2/transport/http"
+	consul "github.com/go-kratos/kratos/contrib/registry/consul/v2"
 
 	_ "go.uber.org/automaxprocs"
+
+	"github.com/hashicorp/consul/api"
+	"github.com/google/uuid"
 )
 
 // go build -ldflags "-X main.Version=x.y.z"
 var (
 	// Name is the name of the compiled software.
-	Name string
+	Name string = "hello-service"
 	// Version is the version of the compiled software.
-	Version string
+	Version string = "0.0.1"
 	// flagconf is the config flag.
 	flagconf string
 
-	id, _ = os.Hostname()
+	id = fmt.Sprintf("%s-%s", Name, uuid.New().String())
 )
 
 func init() {
@@ -34,6 +39,15 @@ func init() {
 }
 
 func newApp(logger log.Logger, gs *grpc.Server, hs *http.Server) *kratos.App {
+	//kratos集成的consul: 微服务退出时 会自动退出
+	// new consul client
+	client, err := api.NewClient(api.DefaultConfig())
+	if err != nil {
+		panic(err)
+	}
+	// new reg with consul client
+	reg := consul.New(client)
+
 	return kratos.New(
 		kratos.ID(id),
 		kratos.Name(Name),
@@ -44,6 +58,8 @@ func newApp(logger log.Logger, gs *grpc.Server, hs *http.Server) *kratos.App {
 			gs,
 			hs,
 		),
+		// with registrar
+		kratos.Registrar(reg),
 	)
 }
 
@@ -74,7 +90,7 @@ func main() {
 		panic(err)
 	}
 
-	app, cleanup, err := wireApp(bc.Server, bc.Data, logger)
+	app, cleanup, err := wireApp(bc.Server, bc.Data, &bc, bc.Jwt,logger)
 	if err != nil {
 		panic(err)
 	}
